requirements.txt
alembic==1.18.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
docker==7.1.0
Flask==3.1.2
Flask-Migrate==4.1.0
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.1
gunicorn==25.0.1
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
python-dotenv==1.2.1
requests==2.32.5
simple-websocket==1.1.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wsproto==1.3.2
WTForms==3.2.1
paramiko==4.0.0
PyYAML==6.0.3
